requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
captcha==0.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==3.2.19
django-ckeditor==6.5.1
django-debug-toolbar==4.1.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
filesig==0.3.0.post1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nexmo==2.5.2
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
twilio==8.2.1
typing-extensions==4.5.0
urllib3==2.0.2
wrapt==1.15.0
yarl==1.9.2
